requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
inflection
flask-serialize
